// src/app/api/model/addressRequest.ts
/**
 * WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddressRequest { 
    city?: string | null;
    region?: string | null;
    postalCode?: string | null;
    country?: string | null;
    street?: string | null;
    apartment?: string | null;
}